requirements.txt
# requirements.txt

# --- Interfaz y Chat ---
streamlit
streamlit-chat

# --- LangChain ---
langchain>=0.2.5
langchain-community>=0.2.5
langchain-openai>=0.1.8 # <--- AÑADIDO

# --- Integraciones específicas ---
langchain-chroma>=0.1.1
chromadb>=0.5.0

# --- Dependencias para funcionalidades ---
pypdf
lark
python-dotenv